server/requirements.txt
openenv-core>=0.2.3
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
openai>=1.0.0
